+++ pkg/orm/model.go
@@ -6,7 +6,7 @@ import (
 )
 
 type Model struct {
-	ID        uint32 `gorm:"primarykey"`
+	ID        int64 `gorm:"primarykey"`
 	CreatedAt time.Time
 	UpdatedAt time.Time
 	DeletedAt gorm.DeletedAt `gorm:"index"`
